backend/requirements.txt
alembic==1.17.1
bcrypt==5.0.0
blinker==1.9.0
certifi==2025.11.12
click==8.3.0
cloudinary==1.44.1
colorama==0.4.6
Flask==3.1.2
Flask-Bcrypt==1.0.1
flask-cors==6.0.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
PyJWT==2.10.1
PyMySQL==1.1.2
python-dotenv==1.2.1
six==1.17.0
SQLAlchemy==2.0.44
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
